requirements.txt
Django==4.2.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.16